--- README.md
@@ -1,13 +1,0 @@
-# aralikesdjango
-
-# django
-pip install django
-
-# django rest framework
-pip install djangorestframework
-
-# image upload module
-pip install pillow
-
-# mysql jsonfield module
-pip install jsonfield
