+++ notebook/api/urls.py
@@ -3,11 +3,11 @@ from django.conf import settings
 from django.conf.urls.static import static
 from . import views
 
-app_name = 'api'
+# app_name = 'api'
 
 urlpatterns = [
-    path('delete/', views.deleteView, name='delete'),
-    path('insert/', views.insertView, name='insert'),
+    path('', views.upload),
+    path('delete/', views.deleteView),
 ]
 
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
